enh: reworked io routines
All io routines have been reworked. They now have a hierachy and
is more stable in the sense that each Sile is inherited through the
overlying code that owns the data file type.
With this methodology one can have the same file-ending and decisevely
return the correct file depending on the code.

Signed-off-by: Nick Papior <nickpapior@gmail.com>