enh: moved arguments to specific class handlers
Moved arguments to be handled by the parenting class
that enables the constructs.

For instance the entire sgeom script was focused on the
Geometry class.
Instead of forcing the options and handling in the script
(sgeom), we may change the actions to perform the required
actions dependent on the input sequence.

The constructs are very basic:

1. Each class creates arguments with appropriate actions.

2. Any top-level script or other class may inherit these
   arguments to create even more complicated constructs
   and control the arguments without ever complicating the
   script itself.

3. Outside applications may also "steal" arguments to control
   the sequence appropriately.

This is apparent by looking at the updated sgeom script.
Now it is only 20 or so lines of code, _everything_ is
handled in the Geometry class.

These changes opted for improving and creating consistent
interfaces for common algorithms. See e.g. the new directory
  utils

Signed-off-by: Nick Papior <nickpapior@gmail.com>