enh: added sdata as an initial code
Due to the previous enhancement we can collect all
command line options into a single command line utility.

This is named
  sdata

It creates the command line arguments dependent on the
file it is passed.

As such it is exactly the same as sgeom for files which
only has geometry information (xyz, XV, etc.)

Siles which does not have implemented an ArgumentParser
will issue a warning about missing implementation.
This is on purpose and any change in the corresponding
Sile's will result in immediate benefit in the sdata
command.

Signed-off-by: Nick Papior <nickpapior@gmail.com>