enh: many features enabled for the tbtrans files, and many fixes
Changed the reciprocal cell to include 2 * pi. For newcomers this
seems like it is much more intuitive.

Updated documentation a couple of places (we need to go through all
of it and abide to:

        Parameters
        ----------
        ia : `list` of `int`
           Atomic indices
        all : `bool = False`
           `False`, return only the first orbital ...

Removed the need for opening/closing NetCDF files in the SileCDF.
This was redundant as the NetCDF file is not a linear file.
The SileCDF also implements a few of shorthands for retrieving values
vs. variables.

The sisl/io/tbtrans.py has been overhauled. Many new routines
for mingling with the orbital currents.
Now one may request direct unit-cell -> supercell bond/orbital currents
for increased flexibility.

Signed-off-by: Nick Papior <nickpapior@gmail.com>