bug: fixed sgeom and the reading of the lattice parameters and geometry
The read_sc and read_geom sections in the fdf file should not be
forced open, we have to re-read the entire file in case the key isn't found.

The sgeom script did take the arguments from the input file (which is not
desirable). Instead we now take the inputfile and read the geometry and
retain the argumenst from that.

tbt.nc fixes:
 - Fixed k-points in tbtrans for Gamma-only calculations
 - Change Arng to Orng to reflect that it is an orbital
   range and not atomic.

Reformatted the xyz file to also contain the nsc fields

Signed-off-by: Nick Papior <nickpapior@gmail.com>