bug: ensured arrays in many cases
In case the user uses ranges to extract sub/remove etc.
it would fail.
Now we automatically cast all indices to arrays.
This has the consequence that quering sc2uc and other functions
will ALWAYS return arrays (even if the input is a single entry).
This is more consistent and I expect this to be better.

TODO go through the code to ensure that indeed everything works as expected

Signed-off-by: Nick Papior <nickpapior@gmail.com>