enh: added operators to Hamiltonian, fixes #14
Adding operators to Hamiltonian enables intrinsic operator
functionality such as:
   NewHamiltonian = Hamiltonian * 2
and
   Hamiltonian *= 2
Currently, are all operations elementwise.

Operators are implemented in __?__ and __i?__ variants
such that in-place operations may occur.

Added tests of the Hamiltonian format and the operators.

Added a copy routine for the Hamiltonian object.

Fixed the floordiv operators check.
Currently we do not check the other div operations as it is
easiest to compare integers

Signed-off-by: Nick Papior <nickpapior@gmail.com>