bug: fixed ranges in utils, negative energies
The ranges in the utility module did not function
exactly like I wanted them. We have now fixed
them so that one may specify [..][..]|..[..]|..
which makes it much more feasible to use.
- Added more tests regarding this

Added a pre-step to the arguments of the scripts
such that negative numbers that may be cast are automatically
prepended a space to ensure they are not interpreted as
options.

Moved the tbtrans output footer up to the top to make
it easier to read.

Fixed a bug with the AtomRange casting them to a
linear array of orbitals, we should use concatenate
instead of casting to array and flattening.

Signed-off-by: Nick Papior <nickpapior@gmail.com>