api: updated construct for recieving a function
I felt the need to change the API for the construct function.
The old construct only worked with atoms with one orbital. Now
it works with any number of orbitals because the user should supply
the function that sets up the tight-binding parameters. It is
much easier to comprehend.

Due to the one-orbital picture is much used there is a variant
by calling the create_construct function.

Signed-off-by: Nick Papior <nickpapior@gmail.com>