enh: added Atoms class to retain information regarding list of Atom's
The old way of storing an array of Atom objects which may have had many
duplicates was inefficient. Now the geometry has a contained Atoms class
associated which essentially contains a list of unique Atom objects and
a list of indices to refer to the unique list. This Atoms object does not
know anything else and it deals with automatic overriding of Atom such that
one can still do:
   Geometry.atom[i] = Atom(...)
this is handy because it automatically detects whether a new specie is necessary.

This fixes #7.

Signed-off-by: Nick Papior <nickpapior@gmail.com>