enh: added Shape directory with commonly used shapes
Their usage is mainly to try and speed up the construction of the
sparse matrices using direct shapes.
The old method of iter_block(...) is very good, but it is performing a
random lookup of the atoms. Instead one can do a fixed grid by moving
rectangles/circles around in a fixed section to ensure the complete
grid is sampled.

This means that the Geometry class has gotten two new methods:
  - within
  - within_sc
which both return the same thing as close, close_sc, respectively.

Instead of a arguments xyz_ia and dR, there is a shapes argument
which contains the center (xyz_ia) and their extend (dR) and thus
are fully controlled from outside.
However, to function properly one _has_ to have an increasing order
of shapes, i.e. shape[0] < shape[1] < shape[...]
etc.

Signed-off-by: Nick Papior <nickpapior@gmail.com>