enh: tbtrans data handler
Added bond-currents info to --info, made true statements have + instead
of - (two ways of seeing whether data resides).

Updated documentation for the tbtnc Sile

Added new keywords for returning the summed DOS for a given set of atoms, the
DOS will be returned in an array of the same length as the queried atoms.
Secondly this returned DOS will _not_ be normalized.

Signed-off-by: Nick Papior <nickpapior@gmail.com>