enh: added atom to rotate functions, and various bug-fixes
Now rotate can take an optional atom argument to only rotate
a subset of atoms. Note that origo is still default [0, 0, 0].
Also, origo may take an integer which corresponds to the atomic
index's coordinate.

Bug-fix for get_sile when a {XYZSile} specification was present.

Added documentation for winSileW90.

Bug-fix for ensure array in self-energy.

Possible speed-up of quaternion due to C-indexing.

Signed-off-by: Nick Papior <nickpapior@gmail.com>