enh: added supercell index specification and supercell fill
In case one only has a 2D system sisl can now figure out the dimensions,
but only when the number of supercells is explicitly set corresponding
to the dimensions. I.e. if

  >>> s = SuperCell(..., [1, 3, 1])
  >>> s._fill_sc((1,))
  (0, 1, 0)

so where 1's are present in the supercell it is filled with zeroes.

Secondly this is used with the Hamiltonian object which now enables:

   >>> H[0, 1, (0, 1, 0)] = ...

which sets the Hamiltonian element between the first orbital in the
unit cell and the second orbital in the (0, 1, 0) supercell.
This makes it easier to set the coupling constants along axis.

This fixes #27.

Signed-off-by: Nick Papior <nickpapior@gmail.com>