enh: added orbocc from ScaleUp
Now the geometry tries to read in the orbocc to correctly set
the number of orbitals.
Secondly, if the geometry used when creating the rham Hamiltonian
does not have the correct orbitals it will try and read the
orbocc file and automatically correct the atomic species.
This should capture the majority of use cases.

Reverted the rham to read the entire file into memory. This sadly
has the effect that we need to parse it twice.

Signed-off-by: Nick Papior <nickpapior@gmail.com>