enh: split the Hamiltonian into a SparseAtom and SparseOrbital class
This is an internal API change to more easily accommodate similar features
from different sparse patterns.
I.e. in this case we want to make it much easier to create density-matrices
and overlap matrices etc.

Now one can easily create an atomic sparse pattern from a Geometry, or
an orbital sparse pattern from a geometry.

This makes the Hamiltonian very easy to implement.

TODO: add a spin-equivalent to handle spin-components (to not re-implement it
      everytime)

Signed-off-by: Nick Papior <nickpapior@gmail.com>