api: created different classes for spin sparse matrices and non-spin matrices
This is mainly to allow more diverse classes, e.g. DensityMatrix, EnergyDensityMatrix
etc.

Currently this is implemented for Hamiltonian and DynamicalMatrix (now default to Hessian).

There shouldn't be any changes for the end-user, albeit there will come more classes now.

The spin handling is now a bit different because the spin is now a separate class contained
as the geometry. This means that one may inquiry the type of spin configuration from the
type Hamiltonian.spin.kind == Spin.UNPOLARIZED
for instance.

Signed-off-by: Nick Papior <nickpapior@gmail.com>