enh: enabled a performance selector
For certain systems it may be faster to use one
algorithm while for others another algorithm may be
faster.

Now sisl implements a Selector class which enables
automatic choice of the most performant one.

Currently it is implemented for the creation of
the Pk routines where several ways of implementing
it may be done.

Signed-off-by: Nick Papior <nickpapior@gmail.com>