enh: added distance function to get a list of distances
The distance method enables obtaining a list of distances to the
neighbouring atoms within separate shells.
I.e. one may do:

   Geometry.distance(<idx>, R=2, tol=0.3)

which returns a unique list of distances (up to 2 angstroms)
with shells of 0.3 angstroms, starting from 0.3 * 0.5.
I.e. the shells are: 0.15, 0.45, 0.75, 1.05, 1.35, 1.65, 1.95.

This is handy for figuring out the distances between atoms.
I.e. if there are defects, or if one is not sure of the bond-distance
of certain atoms.
The routine is rather powerful due to the list input of the tolerances,
i.e. to create shells of different radius. This may be usefull for very
long distance checks. Any periodic array of atoms has a decreasing difference
between different lattice positions, hence a reduced tolerance level may
be required.

Signed-off-by: Nick Papior <nickpapior@gmail.com>