enh: speeded up sub by using array-indexing, i.e. no for-loops
Now the Sparse class implements a generic routine for subbing the
sparse matrix, and or removing rows/columns.
These routines are based on array-slices, so no for-loops are introduced
which should make this the preferable way of creating structures.
Added two tests to use these new routines.

Forgot to add the remove function for SparseOrbital, now it is there.

Changed all where calls with (..).nonzero()[0] which is the equivalent
but faster (more direct call).

Signed-off-by: Nick Papior <nickpapior@gmail.com>