bug: fixed some multi-orbital bugs in tbtrans
Lots of small bug-fixes in the bond-currents. These
should only have been visible when performing calculations
with atoms with more than 1 orbital.

Huge performance boost in the tbtncSile routines for calculating
the necessary bond-currents etc. Now many of the operations are
performed in array notations, rather than looping every element.

Also added preliminary calculation of the current via
the tbtnc sile with default and custom parameters.
This is so far not checked.

Signed-off-by: Nick Papior <nickpapior@gmail.com>