enh: added plotter (not usable in its current state)
And changed the __call__ interface to accept the object with
eigh as a routine. This makes it easier to use because we do
not need to implement the eigh stuff with BrillouinZone objects.

TODO consider changing this to make the BrillouinZone objects contain
the objects where eigh is present. This makes it unique to the class,
but will ensure no wrong calls.

Signed-off-by: Nick Papior <nickpapior@gmail.com>