api: changed the BrillouinZone object for greater flexibility
Now the BrillouinZone object accepts _any_ object which has cell/rcell
as properties. They are used to construct the k->kb values.

This choice also leverages several problems of the BrillouinZone object.
Now any function call on the BrillouinZone object which isn't in the object
gets forwarded to a loop across BrillouinZone k-points and the associated
function call on the retained object. This enables easier access to
internal functions for the parent object.

Signed-off-by: Nick Papior <nickpapior@gmail.com>