maint: added _numpy code to create partials to bypass defaults
Now we can call:

arrayi(..) == array(.., dtype=np.int32)
arrayl(..) == array(.., dtype=np.int64)

which is really easy to use, compared to the much verbose defaults.
This is mainly intended for several of the numpy routines which defaults
integers to np.int64. We could simply use these, but to ensure it will
work in the netcdf files, we will retain the np.int32 for the time being.
In any case, if we are dealing with a 3D perodic structer with nsc=[3,3,3]
we will be "limited" to 79536431 sparse elements.

Signed-off-by: Nick Papior <nickpapior@gmail.com>