enh: added integer electrode specifications
This should be relatively more easy due to not
having to request the case-sensitive electrode names.

Converted all electrode requesting function transfer the argument
to the corresponding electrode string, i.e. if an integer is provided.

Added defaults to the transmission and current calculation. This is
because typically we only deal with 2-terminal devices in which case
it is easy.

Signed-off-by: Nick Papior <nickpapior@gmail.com>