enh: enabled writing a TSHS
Now the binary files can be written by sisl.
This has introduced certain changes and updates to the fortran
interfaces, i.e. when writing the status should be unknown
and action is write.

Changed the way diags works, it now returns a new diagonal matrix
with the same shape as the parent object. One can still reduce
the number of dimensions and the data-type. Tests have been updated.

Signed-off-by: Nick Papior <nickpapior@gmail.com>