enh: added plane for supercell and speeded up close
Now SuperCell class implements a plane method to retrieve
a normal vector and point on the plane (the normal vector ALWAYS
points outwards).

These utilites are now used in the close() method to reduce computations
of unnecessary supercell indices.
This has proven to be twice as fast for bigger systems.

Signed-off-by: Nick Papior <nickpapior@gmail.com>