enh: added sum to the sparse matrix, implemented this for rij
Added Rij routines to return vectors between the sparse elements.
Also added orij and oRij to the geometry.

Bugfixes in tbt.py after converting the _array module.
Also removed the dH files such that they truly are deprecated.
Now dH and dSE are delta files.

Added a sum routine to the SparseCSR class. This currently
implements row sums and last dimension sums.

Simplified the rij routines to not work for what='orbital' if
it is a SparseAtom. Perhaps this may be re-enabled later.

Added more tests for the above and the new sum routine.

Signed-off-by: Nick Papior <nickpapior@gmail.com>