bug: fixed XSF file for vesta, read/write multiple energy points
XSF files for VESTA cannot have blank lines, now we insert # for everything.

The Hamiltonian will now complain if you will try write a Hamiltonian with 0
non-zero elements.

deltanc file could, apparently, not read/write multiple energy-points.
I am not to sure why this was the case, but know I have changed it slightly
and added more tests for this to assert it works.

Signed-off-by: Nick Papior <nickpapior@gmail.com>