enh: added copy to Orbital and SphericalOrbital, major clean-up, add Orbital to Atom
Added copy routines to Orbital and SphericalOrbital classes (not AtomicOrbital yet).

Major clean-up of the orbital class methods.
- Swapped keywords for equal (phi, radial) from (radial, phi).
- added scale function for the Orbital class
  TODO: add to SphericalOrbital
- Added get/set-state to Orbital and SphericalOrbital to allow
  pickable objects. Note that SphericalOrbital has to convert to
  a grid (currently of 1000 points) because functions are not
  necessarily pickable.
- Changed argument order of SphericalOrbital to allow direct interpolation
  functions passed.
  - This also changes set_radial.

Implemented Orbital as base class for orbitals in the Atom object.
Instead of R being an array in each Atom class it is now extracted from
a list of orbitals. I.e. Atom.orbital is a list of Orbital objects.
The initialization of the Atom object has changed, rather dramatic
because it should be able to handle this change. Currently we have:

 - Atom(Z, orbital, .., **kwargs)
  where orbital may be:
     - list of Orbital
     - list of R
  where if floating points are passed it will correspond to
     - list of Orbital(R)
  this should make it backward compatible and make it a bit easier
  to work with.
  For backward compatibility R= still works unless orbital= is used as an
  argument.

All .orbs references have been deprecated, currently I think they were only
used internally. So I guess this is fine. Now .no exists which is equivalent
to the length of the .orbital attribute.

This has produced a large code change due to tests using orbs=...

We are closing in on creating orbitals in real-space grids.

Signed-off-by: Nick Papior <nickpapior@gmail.com>