enh: added grid plotting of phi components
Added __iter__ to the atom (looping over orbitals)

Added a phi routine to the Grid class which takes a vector with
same length as the orbitals in the attached geometry. Then it will
expand the wave-functions from the geometry to the current grid.
Currently this routine has some rather inefficient steps, however
I guess the most time-consuming part is calculating the spherical
coordinates and the phi(..), so I guess it is fine. Perhaps for
really large systems it may be problematic.

Fixed a bug in the xyz2spher.

Fixed a bug in the cube file to correctly get the geometry origo.

Added a toGrid function for an orbital which enables easy calculation
of the single orbital on a grid.

Fixed radial/phi functions to return the data in the same shape as input,
also limited memory usage by reducing temporary arrays.

Fixed bug in SuperCell for origo specification.

Signed-off-by: Nick Papior <nickpapior@gmail.com>