enh: added warnings to the replace function
Added a fill routine for the Grid class (easy overwrite
of values).

The psi function on the Grid *always* removes cross-boundary terms,
however, it still lacks all the neighbouring atoms onto the unit-cell.
So periodic calculations are not correct in its current state.

Enabled reading the Hamiltonian from the fdf file by preferring <>.nc, then
<>.TSHS.

Added a warn if the ChemicalSpeciesAndLabel block was not found in the fdf.

Speeded up calculating the wavefunction on the grid considerably by reducing
calculations to the subset where R is defined.

Fixed a sign-error when comparing against Siesta.

Signed-off-by: Nick Papior <nickpapior@gmail.com>