enh: added supercell wavefunctions, large speedup in psi calculation
Enabled supercell wavefunction calculations. Also enabled k-point wavefunctions.

Many of the loops for the psi calculation could be moved to outside and will thus
be much faster for systems with atoms with many orbitals. I.e. now the indices
are gathered per atom (for the maxium R), and then subsequently calculated for
all orbitals on this atom, finally adding it to the grid.

A bug fix for skewed lattices has been corrected, the R-vector dependent on the indices
did not take off-diagonal cell elements into account.

The loop is now over all supercell atoms which makes it extremely slow for large systems,
however, we quickly skip the atoms that does not have points in the region of the grid
which should be rather quick.

Cleaned up Orbital classes and added a psi_rtp call which accepts spherical coordinates.
This is particularly important for performance in the Grid class because we can
calculate the spherical coordinates per-atom, and then reuse them per-orbital.

A possible bug has been corrected by removing all mentions of origo in calculations.
This is not too thought out. Would the user expect origo to be intrinsic in the coordinates
and operations, or are they merely used for finalization.
Currently the origo is mainly usable for the Grids where one can concentrate the grid
by selecting an appropriate origo.

Bug-fix for the Grid.index function to resemble that of Siesta.

TODO check whether my min_max function actually works. I think we are skipping some elements
in the psi calculation. I need to check (equation of circle to min/max of indices).

Signed-off-by: Nick Papior <nickpapior@gmail.com>