bug: fixed skewed wavefunction calculations
Now sisl will correctly produce the skewed lattices wavefunction.

Instead of "guessing" the min/max values of the indices we create a sphere
with the maximum radius of the atom, then we calculate them exactly on this
sphere. However, the discretization may undercount some points.
This hasn't been tested thoroughly, but we expect the cases to be very few.

Added a spher2cart code to create bothways coordinates.

Added a test for the spherical to cartesian and back.

Signed-off-by: Nick Papior <nickpapior@gmail.com>