enh: speeded up calculation of psi on a grid
Some of the loops have been moved around such that
we calculate the min/max indices on before-hand.
When looping the atoms we simply offset the indices
by the supercell indices.

Also in-lined the idx2R and cart2spher for increased
performance.

Fixed a bug when using real coefficients (was automatically
converted to complex). Now we simply do asarray to retain
the coefficients.

Signed-off-by: Nick Papior <nickpapior@gmail.com>