enh: added info as kwargs to EigenSystem, also added DOS calculator
The EigenSystem now has an extra field containing a dictionary of
additional information provided by the user. This can be useful
in situations where these extra flags is required to be passed around
with the object. E.g. the k-point at which an eigenstate is calculated.

Secondly, added a preliminary DOS calculator, currently we implement
the Lorentzian and Gaussian distributions.

Signed-off-by: Nick Papior <nickpapior@gmail.com>