bug: fixed HSX reading, a first attempt on the reading
The HSX file has now entered the pool of files. It is currently
untested.

Added a couple of routines to the fdf file. I am currently thinking
about a general priority tree to read geometries etc. from the output
files of siesta. This could potentially lead to unexpected behaviour
(from a user perspective), yet I think it makes sense because a relaxation
calculation would then return the relaxed geometry (which is expected).
Possibly we could try and read the method used and subsequently decide
on *what* to return.

Enabled XV files to return the atomic number according to the species index.

Signed-off-by: Nick Papior <nickpapior@gmail.com>