enh: added pyamg code for the Grid
Since I plan on using PyAMG for the Poisson solution I thought it'd
best was to add the pyamg to the class it-self.

This may be rather arbitrarily chosen but it leverages many support functions.

I am not sure whether it should be a sub-module where one should decorate
an object creation or what (to limit non-sisl methods for classes).
However, this may come later.

Added a few tests to create and check the sparse matrices in a csr_matrix
method.

Also updated the way the boundary conditions work. Previously the BC were
handled with 3 BC. However, for Neumann and/or Dirichlet one may envision
lower and upper parts of the box to have separate BC, and as such we need
an abstraction of these 6 possibilities. Note that for periodic BC both
lower and upper *MUST* have that BC.

Signed-off-by: Nick Papior <nickpapior@gmail.com>