enh: finally fixed the last issues related to the iter_block_shape
The iter_block_shape currently only works for Cube/Sphere.
I am thinking about having *any* shape return a minimum rectangle/sphere
(whichever has the largest volume) and iterate that in the sub-space.
However I am not sure how valuable this will be if the regular search
method is faster or of equal performance? Perhaps first when the shapes
are 2D-generalized, infinite cylinders etc.

Added .edge_length to Cuboids to return edge-lengths for the cuboids.

Added an expand routine to PureShape in order to increase the size by a fixed
quantity. For Cuboids, edge-lengths, for Ellipsoids the radius.

Signed-off-by: Nick Papior <nickpapior@gmail.com>