enh: added Grid.index(shape) to retrieve the subset indices, fixes #42
This is similar to the Eigenstate.psi routine by creating a sphere
and find max/min indices on the circumference. Then we create a box
of coordinates.
Lastly we retrieve only the indices within. Note that using this
routine on big shapes may result in extreme memory usage!

Also fixed the _AND toSphere for certain cases.

Added tests to reflect the above.

Signed-off-by: Nick Papior <nickpapior@gmail.com>