enh: enabled reading a DM file.
However, there are many pit-falls. At least the overlap matrix isn't contained
and is currently set to 0 if read from DM file.
When reading from an fdf file it will try and read the geometry and replace if
the # of orbitals is correct. In that case it is usable, otherwise it is hardly
usable.

Signed-off-by: Nick Papior <nickpapior@gmail.com>