enh: got rho working with molecules
I have tested this with un-polarized, polarized and SOC.

It seems it works for SOC, also. But I need Ramon and Roberto
to test this. At least the mulliken fits with the resulting rho.

I still have a few things I don't know is correct, for instance the
DM . spinor. I thikn I have it correct. But...

I have tested this with periodic systems, but can't figure out why I
get wrong charges.

Signed-off-by: Nick Papior <nickpapior@gmail.com>