enh: enabled writing DM/EDM to siesta.DM and siesta.nc files
Now the reading/writing of DM is complete with tests. However
the intrinsic DM file does not contain the overlap matrix, so that
will prevent some things.

Additionally we should probably check that we don't overwrite the overlap matrix
in the *.nc file when issuing write_hamiltonian and write_density_matrix.

Same thing for EnergyDensity matrix for the *.nc file.

Signed-off-by: Nick Papior <nickpapior@gmail.com>