bug: fixed toSphere for the Cuboids, also added preliminary toEllipsoid method
I think using the Ellipsoid will be better because for overlapping orbitals an
ellipsoid will be better at describing a minimal region of space.
Hence we have now a toSphere and toEllipsoid, where the latter should in many
cases produce a shape with smaller volume.

This led to the discovery that Cuboid.toSphere actually created a too small
Sphere. This has been fixed.

Signed-off-by: Nick Papior <nickpapior@gmail.com>