enh: added symmetry to MonkhorstPack
TODO it still has too many points on the line orthogonal
to the symmetry line at k = 0. I.e. we could remove a small
fraction of the points there.
This is to be done later.

Also added an optional requirement of tqdm to display
progress bars.
They are now implemented in:
- construct
- BrillouinZone.as* routines.
  I.e. BZ.asarray().PDOS(eta=True)
  Note how eta is a new keyword!

Signed-off-by: Nick Papior <nickpapior@gmail.com>