enh: re-implementation of rho
Now rho is *MUCH* faster. It isn't as fast as denchar. But at least it
is usable.

It can now correctly create the density for molecules. I need
to check for periodic systems.
However, since I have based it on the wavefunction routine with
within_inf I suspect the periodicity to come "for-free".

This new method is also a bit more memory consuming. But the memory
overhead is much less important than the performance increase.

The easiest performance increase was to sum the upper and lower
triangular part to halve the number of psi evaluations.
This can, as far as I see only be performed for the unit-cell DM.
But this is also the most important one!

Signed-off-by: Nick Papior <nickpapior@gmail.com>