enh: enabled the extraction of the density matrix from TBT.nc files
Using the DM from TBtrans we are now able to plot LDOS(E, r)!
This is extremely easy and may be very useful in further investigations.

This also showed a *bug* in the __*__ math operations of the CSR matrix.
Now the SparseCSR matrix can correctly handle the math operations of
scipy sparse matrices. Before it silently accepted the math commands
without doing anything... Sadly...

Signed-off-by: Nick Papior <nickpapior@gmail.com>