maint: fixed replace function for k-points outside the BZ
When trying to replace a k-point outside the BZ it could
sometimes not find a point. This has been fixed.

Also fixed cases for TRS such that the weights are correct and
accounted for.

Added a staticmethod to the BrillouinZone such that one
may move *any* k-point into the primitive Wigner-Seitz cell.

Signed-off-by: Nick Papior <nickpapior@gmail.com>