enh: enabled Cython Pk setup, fixes #51
Tests indicate that this speeds up the creation by more
than 200 times. However, the main part of computation is still
in the eigh routines.

I have thoroughly tested the polarized setup of Hamiltonians and
also one-time test of NC and SO. I thus suspect it to work as
expected (but more tests are of course needed).

The test_hamiltonian has been updated to not do Hk several times.
There may be remnants for these tests that eventually should be removed.

Changed a few Cython routines to return copies. I think this may be more
memory efficient because then there are no references to too long arrays.
I haven't checked this.

Signed-off-by: Nick Papior <nickpapior@gmail.com>