enh: enabled the calculation of the effective mass tensor from the bandstructure
It seems to work, but I have no real idea about checking it.
For graphene TB it gives numbers around 1 which seems reasonable (for some
k-points). So a better check requires better insight.

This also streamlines the dk matrices by a common prefactor of the phases
and the iRs. They are now unified in a single array when passed.

Signed-off-by: Nick Papior <nickpapior@gmail.com>