bug: fixed EDM shift in siesta_nc reading, fixed orthogonal check in TSHS reads
When reading the EDM it should be shifted by Ef * DM to account for the
Fermi level. This is now implemented.

The TSHS read of Hamiltonian now checks whether the basis is orthogonal (like siesta_nc).

Added shift function to the EnergyDensityMatrix to enable shift according
to a passed density matrix.

Signed-off-by: Nick Papior <nickpapior@gmail.com>