enh: enabled reading supercell Hessian matrices, add make_hermitian
Added a new routine to all sparse matrices (make_hermitian)
which takes all elements and ensures that the matrix is Hermitian.

Now fdf(FC) files may be read for Hessian matrix creations.
It does this by a simple read and set routine which
basically does something like the cut routine.

All FC elements are added and subsequently the new make_hermitian
routine is used to ensure the matrix's Hermiticity.

Implemented the binary search method in the index_sorted, this should
speed up a few things.

Signed-off-by: Nick Papior <nickpapior@gmail.com>