enh: changed all Hessian to DynamicalMatrix
In fact what we get from the read_hessian was always
the dynamical matrix. The Hessian should probably be the
average FC along the displacements.

I have now added a phonon.py similarly to electron.py.

Currently it can do:

 - DOS
 - PDOS
 - velocity (see phonopy)

TODO displacement

Also changed hessianSileGulp to fcSileGulp which is better
reflecting the content of the file.

Signed-off-by: Nick Papior <nickpapior@gmail.com>