enh: made sisl writes to Siesta files compliant with siesta supercell
The siesta files was simply writing the content *as-is* which could produce
siesta problems when using a non-conforming supercell.

I.e. now when a sisl object is written to a siesta sile the sparse matrix
gets converted into the siesta default sparse pattern.
And when they are read, they are converted to sisl pattern.

This is somewhat untested but the tests pass.
Added a test for the dynamical matrix.

Bugs fixed:
  - read-dynamical-matrix now checks units
  - read-hamiltonian now checks units
  - read-force now really does read the forces

Signed-off-by: Nick Papior <nickpapior@gmail.com>