enh: changed the behavior of kavg argument
Now kavg only accepts a bool or a specific integer.
When given a bool, it will k-average (true) or return the k-resolved
quantity (without k-weights).

For an integer it will return the quantity for the given k-index, without
weight.

The kavg=[] will no longer work since it complicates things to heavily.

Signed-off-by: Nick Papior <nickpapior@gmail.com>