enh: reverted 412c0bd0 and improved inner/norm2 handling
Now ElectronState also implements correct overlap matrix handling
for inner/norm2 calculations.
This means we can correctly calculate off-diagonal stuff etc.

Note that inner fails if right is given and it contains
an overlap matrix. In those cases one should probably convert
to the Lowdin basis, and then do the inner-product. However, currently
we don't do that as I am not fully sure about any implications.

Added has_noncolinear to the Spin class. This works as a test for
  is_noncolinear || is_spinorbit
which is the requirement for doubling the basis elements space.

Streamlined inner products with non-colinear states. Actually the
code was rather complicated for doing nothing. Now it is simpler.

Signed-off-by: Nick Papior <nickpapior@gmail.com>