enh: stream-lined spin-orbit and non-colinear matrices
In Siesta the nomenclature is that the spin-boxes have
[[a, b], [c, d]] contained with Im(b) stored as the
negative value.
In sisl this is somewhat counter-intuitive since
we shouldn't propose anything but having positive
signs.

This means that *everything* that is read by sisl
should adhere to the convention outlined
in sisl/io/siesta/_help.py.

This should fix #125 once and for all.

Signed-off-by: Nick Papior <nickpapior@gmail.com>