enh: performance increase for RealSpace methods
Now we use the PEXSI-Sigma notation for the coupling self-energy
calculation. This should always be faster but it is not
clear to me which method is numerically better, I see
differences on the order of 1e-8 eV depending on system sizes.

Since it is faster (and should be equivalent) I now use
the Dyson equation for evaluating the coupling self-energy.

Signed-off-by: Nick Papior <nickpapior@gmail.com>