enh: enabled better reading of matrices from fdf file
When reading from the fdf file it is better if one can expect the
geometry to be correct, i.e. number of orbitals and atomic numbers.
Now the correct geometry is tried to be read through the fdf in combination
with the TSHS file when requesting matrices, this ensures correct orbital
count and atomic numbers.

Added tests for spin-orbit DM/TSHS for Siesta output.

Signed-off-by: Nick Papior <nickpapior@gmail.com>