enh: enabled TSDE writing of DM+EDM, fixes #138
Enabled writing TSDE files with and without Ef arguments.

Cleaned siesta/_src fortran files. Now all unit conversions
are done in the fortran codes and should make it
easier for future developments. I.e. *any* development
should return the correct unit.
The only exception here is that Grid files does not implicitly
know which unit they are in. This is based on the file extension
and is thus not done in fortran.

Removed finalize on input matrices to write files. This is probably
better since this allows users to expect the same sparse matrix after
method call.
Additionally I have changed the finalize(True) to finalize(False)
on sparse matrices in write_* since they are not necessary.
I should probably ensure this in TBtrans for performance reasons.

Added tests for reading/writing DM/TSDE files.

Fixed bug in energydensitymatrix when shifting the EDM.

Signed-off-by: Nick Papior <nickpapior@gmail.com>