enh: enabled reading the fermi-level from the siesta output
I haven't fully decided what name this should have.
Currently it is named read_fermi_level which is probably
justified because it is used extensively. However, it could
be argued that this should be read through read_energy(fermi_level=True)
or something like this.

For now I'll keep this.

It currently allows reading the fermi-level from the TSDE, nc files.

This fixes #126.

Signed-off-by: Nick Papior <nickpapior@gmail.com>