enh: added Berry curvature calculation
Added calculation of Berry curvature for states.
This may be used to calculate conductivity (to be done).
For now it calculates the Berry curvature using:

X. Wang, J. R. Yates, I. Souza, D. Vanderbilt,
"Ab initio calculation of the anomalous Hall conductivity by Wannier interpolation", PRB, *74*, 195118 (2006)

and is thus limited by the number of unoccupied states.
The more un-occupied states the better. However, since
the Berry curvature is scaled with the eigenvalue difference
it truncates relatively fast.

Note: this is untested and requires some benchmarks.
Signed-off-by: Nick Papior <nickpapior@gmail.com>