Fix crash in Overlap.fromsp due to overridden init
Using Overlap.fromsp would crash with the message "__init__() takes from
2 to 4 positional arguments but 5 were given". This was due to the class
instantiation call in super().fromsp which used positional arguments
that coincided with that class's kwargs, but not with the Overlap
object's kwargs. Putting the `dim` kwarg into **kwargs was also
necessary to allow it being passed (but still overridden), similar to
the `orthogonal` keyword.